pyproject.toml
[tool.poetry]
name = "diplom_tasktracker"
version = "0.1.0"
description = ""
authors = ["StanisLove <sta-tun-slav@yandex.ru>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.3"
SQLAlchemy = "^2.0.5"
phonenumbers = "^8.13.7"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.5"

[tool.poetry.scripts]
server = "src.main:run"
setup_db = "src.db:create_all_tables"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
